requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.2.0
cchardet==2.1.7
certifi==2022.5.18.1
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
ffmpeg-python==0.2.0
flash==1.0.3
Flask==2.1.2
Flask-Discord==0.1.69
frozenlist==1.3.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
http3==0.6.7
httpcore==0.15.0
httpx==0.23.0
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio-ffmpeg==0.4.7
instagram-private-api==1.6.0.0
ipapi==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsom==0.0.5
keyboard==0.13.5
lark-parser==0.12.0
lxml==4.9.0
MarkupSafe==2.1.1
mouse==0.7.1
multidict==6.0.2
numpy==1.23.1
oauthlib==3.2.0
orjson==3.7.8
Pillow==9.2.0
prettytable==0.7.2
py-cord==2.0.0
pycares==4.2.1
pycord==0.1.1
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyreadline==2.1
pyreadline3==3.4.1
python-dateutil==2.8.2
requests==2.24.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
roblox==2.0.0
robloxapi==4.7
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
typing_extensions==4.2.0
urllib3==1.25.11
Werkzeug==2.1.2
yarl==1.7.2